Renamed folder to correct type
